projectiles

+++ tankbattle/client/stub/Tank.java
@@ -56,8 +56,8 @@ public class Tank {
                     JSONObject projectile = tank_projectiles.getJSONObject(k);
                     Projectile p = new Projectile();
                     p.id = projectile.getString("id");
-                    p.position.x = projectile.getJSONArray("position").getInt(0);
-                    p.position.y = projectile.getJSONArray("position").getInt(1);
+                    p.position.x = projectile.getJSONArray("position").getDouble(0);
+                    p.position.y = projectile.getJSONArray("position").getDouble(1);
                     p.direction = projectile.getDouble("direction");
                     p.speed = projectile.getInt("speed");
                     p.range = projectile.getDouble("range");
